requirements.txt
pandas
torch
transformers
huggingface_hub
pyyaml
tqdm
